Report컨트롤러와 Request컨트롤러 구조 변경
